Adding API to get game modes available
The code was assuming that all players would have the competitive and quickplay game modes in the dictionary.
This was not the case and there was not a way to figure out which game modes were available for the player without
looking at the raw data.

This change add a new API modes() to get an array of game modes for a player. It also modifying the model for the player
data by adding a 'modes' key to the top level of the model. Now the 'competitive' and 'quickplay' key are part of this dictionary.